+++ resolwe/permissions/tests/test_data.py
@@ -210,8 +210,9 @@ class DataTestCase(ResolweAPITestCase):
         self.assertKeys(resp.data, ['slug', 'name', 'created', 'modified', 'contributor', 'started', 'finished',
                                     'checksum', 'status', 'process', 'process_progress', 'process_rc', 'process_info',
                                     'process_warning', 'process_error', 'input', 'output', 'process_type',
-                                    'descriptor_schema', 'descriptor', 'id', 'process_name', 'process_input_schema',
-                                    'process_output_schema', 'current_user_permissions', 'descriptor_dirty', 'tags'])
+                                    'descriptor_schema', 'descriptor', 'id', 'process_slug', 'process_name',
+                                    'process_input_schema', 'process_output_schema',
+                                    'current_user_permissions', 'descriptor_dirty', 'tags'])
 
         # user w/ public permissions
         resp = self._get_detail(1, self.user2)
@@ -219,8 +220,9 @@ class DataTestCase(ResolweAPITestCase):
         self.assertKeys(resp.data, ['slug', 'name', 'created', 'modified', 'contributor', 'started', 'finished',
                                     'checksum', 'status', 'process', 'process_progress', 'process_rc', 'process_info',
                                     'process_warning', 'process_error', 'input', 'output', 'process_type',
-                                    'descriptor_schema', 'descriptor', 'id', 'process_name', 'process_input_schema',
-                                    'process_output_schema', 'current_user_permissions', 'descriptor_dirty', 'tags'])
+                                    'descriptor_schema', 'descriptor', 'id', 'process_slug', 'process_name',
+                                    'process_input_schema', 'process_output_schema',
+                                    'current_user_permissions', 'descriptor_dirty', 'tags'])
 
     def test_get_detail_no_perms(self):
         # public user w/o permissions