+++ docs/CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 -----
 - Add ``sort`` argument to ``assertFile`` and ``assertFiles`` methods in
   ``ProcessTestCase`` to sort file lines before asserting the content
+- Add ``process_slug`` field to ``DataSerializer``
 
 Fixed
 -----